BadPS4 Testing
Tested and works with real controllers
